Add files via upload
This is an open source HTML, CSS website code that I made as part of my mini projects series. 
When opened it contains five squares, the most important one being the yellow one in the middle which along with the other squares get smaller as they keep on spinning. The other squares that act as blades of this spinning square fan are my favorite colors in order from Blue, Violet, Pink and Green, now the reason I decided to use yellow as the centerpoint color for this project is because it recently grew on me and might even overtake blue as my favorite color soon.

In order to open and use this code. There are two simple steps.
Step 1-If you just want to open the code then just double click it with your cursor since it's just an HTML file which is supported by any browser that you have on your computer.
Step 2- If you want to utilize/improve upon the code for yourseld then please download the HTML file, then right click it with your mouse. Next select the open with function, you may choose whatever IDE that you are comfortable with such as Pycharm, Webstorm or even VSCode; for my case I specifically used VSCode when I made this project.

Enjoy, feel free to download and improve upon my work, but before forking or downloading it please give me some credits. Thank you and have a great day ahead fellow developers.